fix or silence a wide range of different compiler warnings (msvc)

+++ example/naive_monte_carlo_example.cpp
@@ -29,7 +29,7 @@ void display_progress(double progress,
     int barWidth = 70;
 
     std::cout << "[";
-    int pos = barWidth * progress;
+    int pos = static_cast<int>(barWidth * progress);
     for (int i = 0; i < barWidth; ++i) {
         if (i < pos) std::cout << "=";
         else if (i == pos) std::cout << ">";
